Add a cancel method to QuartzTableManager(), and ensure it is called if any errors occur while modifying the quartz database - this prevents partial modifications being applied.
QuartzTableManager::set_revision_number() and apply() now return
void - exceptions are thrown if errors occur.

QuartzBufferedTables now free the memory they're using as soon as
possible when writing changes to disk.

Added QuartzBufferedTable::write() method - when implemented, this
will write the changes so far to disk and free the memory used,
but not make a new revision: this is wanted for some backends where
no sort of the modifications is needed, so they don't need to be
held in memory.


git-svn-id: https://svn-dav.xapian.org:8443/xapian/trunk@2746 6b2f1b92-a4f3-0310-943f-de3cbe41748b