Implemented AllTermsIterator for quartz.  Interaction with QuartzPostList needs to be cleaned up.
Added new functions [un]pack_string_preserving_sort(), which
pack a string in a way which preserves sort order.  Added a
testcase to quartztest.

Now use [un]pack_string_preserving_order() on postlist keys, so
that we can get at terms in sorted order.  (And get_tname_from_key()
added to separate out the term-unpacking as well as the packing)

Enabled test_allterms1 in apitest, but put it in a group of
tests only run for inmemory and quartz.  This should be extended
to other backends.


git-svn-id: https://svn-dav.xapian.org:8443/xapian/trunk@2939 6b2f1b92-a4f3-0310-943f-de3cbe41748b