so instead of everytime we calculate alignment (very expensive operation), we save the result when first time the database is built. (read_mode is false).
every subsequent calls, diff result is read from a database, and we simply
simulate the mapping algorithm again with the diff results. If no database is
specified, then we simply recalcuate the diff output again dynamically.

this makes the html comparator much faster and possible to be used
as a cgi script


git-svn-id: https://svn-dav.xapian.org:8443/xapian/trunk@3228 6b2f1b92-a4f3-0310-943f-de3cbe41748b