database files can now be opened in two modes.. one in read-only mode and one in create read-write mode.
this makes it easier to have database files in read-only mode after
they are built via cvsbuild

side effect:

this means if the first call to a database is a get, then the database
must already exist and it will be in read only mode. any put calls will fail.

however if the first call to a database is a put, then subsequent get calls
are still ok.

this is a limitation, since we usually first build database files (use a lot of
put command), and then do query (do a lot of get command), the above
limitation doesn't apply.


git-svn-id: https://svn-dav.xapian.org:8443/xapian/trunk@3230 6b2f1b92-a4f3-0310-943f-de3cbe41748b