fixed a bug for lines_db..
previously an assumption I used in lines_db is that every line in a file
is associated with at least one commit.. and hence at least one revision.
therefore, before say if I try to query for line 23 of file 3,
if no revisions found, I stop looking at file 3 and start to look at
file 4.

I don't believe this is an error in the mapping algorithm, because it is
possible to miss lines esp we are using diff with -b option.


git-svn-id: https://svn-dav.xapian.org:8443/xapian/trunk@3514 6b2f1b92-a4f3-0310-943f-de3cbe41748b