2 changes in one commit, tsk tsk!
Cause index nasty to go into an infinite retry if it can't lock the DB, instead of dieing.  This is good for ME because I usually feed it stdin.

Also caused passing just the unique term as the nasty document to delete the
document.

e.g.

*Q:id=23
~=Story

will create a record which can be deleted with

*Q=23


git-svn-id: https://svn-dav.xapian.org:8443/xapian/trunk@3823 6b2f1b92-a4f3-0310-943f-de3cbe41748b