Fixed Oranges sort problem which turned out to be a badly described collapse problem; on account of OmMSet of course containing COPIES of the same OmMSetItems that are in collapse_tab but NOT being references to the same instance! So fat lot of good me updating collapse_count in the collapse_tab copy then!
Also added some missing include files (xapian/*.h) to .spec file as they
are needed right now.

And yes, this means the recently reported 9GB sorting bug is NOT reproduced
with a smaller DB, and so yes, Olly, you do seem to have fixed the sorting
bug previously reported and referred to by me this week.

Is it OK to put such chatty comments as CVS commit messages?  Ah well it
will give the data archeologists something to laugh about.


git-svn-id: https://svn-dav.xapian.org:8443/xapian/trunk@4530 6b2f1b92-a4f3-0310-943f-de3cbe41748b